Acousics and articulatory durational measures of Italian and Polish

A dataset with measures from acoustics and articulatory data of Italian and Polish. The dataset contains durational measures of segments, voicing, and tongue gestures.

Usage

dur_ita_pol

Format

A tibble with 1343 observations and 52 variables.

index

observation number within speaker

speaker

speaker's ID

file

audio chunk file name

rec_date

date and time of recording

ipu

SPPAS IPU index

prompt

sentence stimulys

word

word stimulus

time

(s) time of the sentence onset within the concatenated audio file

sentence_ons

(s) onset time of the sentence

sentence_off

(s) offset time of the sentence

word_ons

(s) onset time of the target word

word_off

(s) offsett time of the target word (= C1 onset)

v1_ons

(s) onset time of V1 (= C1 offset)

c2_ons

(s) onset time of C2 (= V1 offset)

v2_ons

(s) onset time of V2 (= C2 offset)

c1_rel

(s) time of C1 release

c2_rel

(s) time of C2 release

voicing_start

(s) time of voicing onset

voicing_end

(s) time of voicing offset

voicing_duration

(ms) duration of voiced interval

voiced_points

number of points out of 5 within the first half of C1 closure in which voicing is present

GONS

(s) onset of C1 closing gesture

max

(s) time of maximum displacement of C1 closing gesture

NOFF

(s) offset of C1 gesture nucleus

NONS

(s) onset of C1 gesture nucleus

peak1

(s) first tongue velocity peak

peak2

(s) second tongue velocity peak

c1_duration

(ms) duration of C1

c1_clos_duration

(ms) duration of C1 closure

c1_vot

(ms) C1 Voice Onset Time

c1_rvofft

(ms) C1 release to V1 offset time

v1_duration

(ms) duration of V1

c1_duration

(ms) duration of C1

c2_clos_duration

(ms) duration of C2 closure

v2_duration

(ms) duration of V2

v_v

(ms) V1 onset to V2 onset (Vowel-to-Vowel) duration

word_duration

(ms) duration of the word

sentence_duration

(s) duration of sentence

language

speaker's native language (Italian, Polish)

gender

speaker's sex (f, m)

glottocode

language Glottocode

item

word ID number

ipa

IPA transcription of the word

c1

first consonant (C1)

c1_phonation

voicing of C1 (voiceless, voiced)

vowel

V1 and V2 (a, o, u)

anteropost

backness of the vowel (back, central)

height

height of the vowel (high, mid, low)

c2

second consonant (C2)

c2_phonation

voicing of C2 (voiceless or voiced)

c2_place

place of C2 (coronal, velar)

speech_rate

speech rate as syllables per second

speech_rate_c

centered speech rate as syllables per second