Data on vowel duration, voicing duration, and vowel height in Italian.

A dataset with durational measures from 19 speakers of Italian.

Usage

ita_egg

Format

A tibble with 3268 observations and 31 variables.

speaker

speaker's ID

ipu

IPU index

stimulus

sentence stimulus

sentence_ons

time of sentence onset in seconds

sentence_off

time of sentence offset in seconds

word_ons

time of word onset (= C1 onset) in seconds

word_off

time of word offset (= V2 ofset) in seconds

v1_ons

time of V1 onset (= C1 offset) in seconds

c2_ons

time of C2 onset (= V1 offset) in seconds

v2_ons

time V2 onset (= C2 offset) in seconds

voice_ons

time of voice onset in seconds

voice_off

time of voice onset in seconds

c1_rel

time of C1 release in seconds

c2_rel

time of C2 release in seconds

stimulus_id

index of the stimulus sentence

sentence

frame sentence

word

target word

c1

C1 (p, t, k)

vowel

V1 (i, e, u, ɛ, a)

c2

C2 (p, t, k)

backness

height

c1_place

C1 place of articulation

c2_place

C2 place of articulation

v1_duration

duration of V1 in milliseconds

c2_clos_duration

duration of C1 closure in milliseconds

rel_voff

duration of C1 release to V1 offset in milliseconds

sent_duration

sentence duration in seconds

speech_rate

speech rate as number of syllables per second (= 8 / sentece_duration)

speech_rate_c

centered speech rate (= speech_rate - mean(speech_rate))

voice_duration

duration of the voiced interval of V1 in milliseconds

vot

duration of Voice Onset Time in milliseconds

voi_clo

duration of voicing during C2 closure in milliseconds

rel_rel

duration of C1 release to C2 release